+++ CHANGELOG.rst
@@ -3,6 +3,7 @@ dev:
       added methods:
         - dump_db: wrapper around ``db.dump`` server method.
         - restore_db: wrapper around ``db.restore`` server methods.
+    - ``openerp_proxy.ext.repr.HField`` added ``F()`` method, which allows to create child field instance
 
 0.6:
     - *Backward incompatible*: Changed session file format.